refactor(popover): split utils into modules
